backend/requirements.txt
# Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1

# Database
sqlalchemy[asyncio]==2.0.37
asyncpg==0.30.0
alembic==1.13.1

# Validation & Settings
pydantic==2.10.6
pydantic-settings==2.7.1
email-validator==2.1.0.post1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.26.0

# Google APIs
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Utilities
python-dateutil==2.8.2

# Background Tasks
apscheduler==3.10.4

# Encryption
cryptography==42.0.2

# Direct bcrypt (for password hashing)
bcrypt==4.1.2

# SendGrid (Email delivery)
sendgrid==6.11.0
